+++ juriscraper/opinions/united_states/state/nd.py
@@ -52,6 +52,11 @@ class Site(OpinionSiteLinear):
         seen_urls = set()
 
         for row in self.html.xpath('//table//div[@class="row"]'):
+            onclick = row.xpath(".//button[@onclick]/@onclick")
+            if not onclick:
+                logger.info("Skipping row without a 'View Opinion' button")
+                continue
+
             raw_name, *values = list(
                 map(str.strip, row.xpath("./div[1]/p[1]/text()"))
             )
@@ -59,7 +64,7 @@ class Site(OpinionSiteLinear):
             # Do the URL check
             url = urljoin(
                 self.base_url,
-                row.xpath(".//button[@onclick]/@onclick")[0].split("'")[1],
+                onclick[0].split("'")[1],
             )
             if url in seen_urls:
                 logger.info(